+++ lib/gps_camera_snap.dart
@@ -3,3 +3,4 @@ library;
 export 'src/gps_camera.dart';
 export 'src/models/gps_camera_config.dart';
 export 'src/utils/maptiler_validator.dart';
+export 'src/utils/map_tile_provider.dart';